app/requirements.txt
Django==3.2
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt
djoser==2.0.3
idna==2.10
oauthlib==3.1.0
psycopg2-binary
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests-oauthlib==1.3.0
six==1.15.0
django-utils-six==2.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3
django_otp
django-two-factor-auth==1.16.0
django-phonenumbers==1.0.1
Pillow==10.2.0
drf-yasg